Create run.py
This is an example run script, you will need to modify the image import code to import whatever images you plan on using. Other than that this should be everything you need to run the DanGAN

I've left in the code that saves batches of images every 64 iterations as well as code to make some charts, but disable these if you don't want to fill up your hard drive with GAN images.